server/requirements.txt
django
psycopg2-binary
gunicorn
djangorestframework
djangorestframework-simplejwt
celery
redis
numpy
scikit-learn
requests
django-cors-headers
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2